requirements.txt
tensorflow==1.13.1
Keras==2.2.4
flask==1.1.1
matplotlib==3.1.2
numpy==1.17.4
pandas==0.25.3
pdfminer3==2018.12.3.0
pymongo==3.9.0
spacy>=2.2.0,<3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#egg=en_core_web_sm
dnspython==1.16.0
python-docx==0.8.10
Pillow==7.0.0